Simplify insert
Reading the wikipedia article on AVL more closely, it leverages the fact that
the rebalance operations are guaranteed to absorb an inbalance due to insertion.
In the rot_lf and rot_rt cases, we can only be out-of-whack because one child
was taller and the rebalance places that child where its height won't matter.
The rot_lf_rt and rot_rt_lf cases always shrink the tree, for both insertions
and deletions.