Make a 'destructive' version of join
By destructive we mean, move into/from the existing maps rather than creating a
new map.  This may avoid clones and ref-count updates for those nodes that are
solely owned by the argument maps.  When the nodes are shared, they will be
cloned.

For convenience, we left a version that takes borrowed maps and constructs a new
map as before.  In general, a user can get the same behavior by calling clone
before calling the version that consumes the maps.