Deserialize more efficiently
Previously, we created minimially sized leaves and merged them as long as
possible.  This may require a large number of memory moves.

Now, we try to start with fat leaves, to avoid most of the memory moves at the
leaf level.  Since most elements are at the leaf level, this should
significantly reduce the number of memory moves during map deserialization.